Add iBeacon advertisement into library
First implementation of iBeacon advertisements into the existing
Eddystone advertising library.  This was done to leverage the existing
structure of the Eddystone advertising library, making it fairly easy to
incorporate iBeacon advertisements for Chrome OS.

The main difference between the Eddystone advertisements (URL and UID)
and iBeacon advertisements is that Eddystone uses a service advertisement
whereas iBeacon uses a manufacturer advertisement.  After digging into
the Chrome OS BLE advertising API it was discovered that you can pass in
manufacturer data as well as service data, making iBeacon transmission
possible.

One interesting discovery was that the BLE manufacturer advertisement
from Chrome OS doesn't include the flag bytes that prepend the
manufacturer data in the packet.  This was discovered using a BLE
sniffer.

In addition, it was found that the Bluetooth MAC address is rotated
every time a new advertisement is registered with the OS, making it
difficult to link a BLE advertisement back to the hardware.