Use numbers to input major/minor instead of hex
To be consistent with the iBeacon standard, which classifies major and
minor values as integers not hex strings.

In the initial implementation of iBeacon in the Eddystone advertising
library, major and minor were inputted as hex strings (2 bytes in
length).  This was done for simplicity as the Eddystone library already
used hex strings for UID namespace and instance identifiers.

Now the iBeacon advertisement takes the major and minor at numbers and
validates them before converting them into bytes for the advertisement.